+++ src/components/status/index.tsx
@@ -21,6 +21,7 @@ export default function Status({
         status?.toLocaleLowerCase().includes("resolved") ||
         status?.toLocaleLowerCase().includes("success") ||
         status?.toLocaleLowerCase().includes("complete") ||
+        status?.toLocaleLowerCase().includes("occupied") ||
         booleanVal ? (
         <span className=" p-1 px-3 bg-green-500/15 text-green-500 rounded-full">
           {truthyMessage || status}