+++ src/components/tables/reports/daily-room.tsx
@@ -1,27 +1,55 @@
-import { useState } from "react";
+import { useCallback, useState } from "react";
 import { apartmentById } from "../../../types/apiData/apartment";
 import Filter from "../../filterAndSort/filter";
 import Select from "../../inputs/select";
 import Search from "../../inputs/search";
 import LoadingButton from "../../button";
 import ExportSelect from "../../inputs/select/exportSelect";
-import useGetAllReportsLists from "../../../services-hooks/useGetReportList";
 import { revenueReportList } from "../../../types/apiData/reports";
 import NoResult from "../../noResult";
 import Pagination from "../../pagination";
+import formatDate from "../../../utils/isoDateConverter";
+import useGetRevenueReport from "../../../services-hooks/reports/revenue";
+import useGetReportSummary from "../../../services-hooks/reports/report-summary";
 
 export default function DailyRoomReportTable() {
-  const [apartment, setApartment] = useState<apartmentById>({} as any);
   const [type, setType] = useState("");
+
+  const [filterDates, setFilterDates] = useState<{
+    start_date: string;
+    end_date: string;
+  }>();
+  const [apartment, setApartment] = useState<apartmentById>({} as any);
   const [currentPage, setCurrentPage] = useState(1);
+
   const { data, isLoading, isFailed, setIsFailed, retryFunction, pagination } =
-    useGetAllReportsLists({ page: currentPage, type: "daily-room" });
+    useGetRevenueReport({
+      page: currentPage,
+      apartmentId: String(apartment?.id || ""),
+      start_date: filterDates?.start_date,
+      end_date: filterDates?.end_date,
+      group: "day",
+    });
+
+  const { data: reportSummary } = useGetReportSummary({
+    apartmentId: String(apartment?.id || ""),
+    start_date: filterDates?.start_date,
+    end_date: filterDates?.end_date,
+  });
+
+  const handleCustomersFiltering = useCallback(
+    (start_date: string, end_date: string) => {
+      setFilterDates({ start_date, end_date });
+    },
+    []
+  );
 
   return (
     <div className="w-full flex flex-col gap-5">
       <div className="flex items-center justify-between flex-col md:flex-row gap-4 p-4 rounded-md border flex-wrap lg:flex-nowrap">
         <div>
-          <Filter />
+          {" "}
+          <Filter actionHandler={handleCustomersFiltering} />
         </div>
         <div>
           <Select
@@ -52,9 +80,8 @@ export default function DailyRoomReportTable() {
           <ExportSelect id="report" />
         </div>
       </div>
-      {/* table */}{" "}
       <div className="w-full rounded-lg border p-5 flex flex-col gap-5 overflow-auto ">
-        {data && data.length > 0 ? (
+        {data && data.length > 0 && apartment?.id ? (
           <>
             <table className=" w-full text-xs overflow-x-auto">
               <thead className="">
@@ -76,34 +103,33 @@ export default function DailyRoomReportTable() {
                 </tr>
               </thead>
               <tbody className="">
-                {data.map((request: revenueReportList) => {
+                {data.map((request: revenueReportList, index: number) => {
                   return (
-                    <tr key={request?.id} className=" border-b">
-                      <td>{request?.date}</td>
-                      <td>{request?.roomSold}</td>
-                      <td>{request?.nightBook}</td>
-                      <td>{request?.totalBooking}</td>
-                      <td>{request?.occupancy}</td>
-                      <td>{request?.ibeRevenue}</td>
-                      <td>{request?.otaRevenue}</td>
-                      <td>{request?.refunds}</td>
+                    <tr key={index} className=" border-b">
+                      <td>{formatDate(request?.date)}</td>
+                      <td>{request?.rooms_sold}</td>
+                      <td>{request?.nights_booked}</td>
+                      <td>***</td>
+                      <td>{request?.occupancy_rate}</td>
+                      <td>{request?.ibe_revenue}</td>
+                      <td>{request?.ota_revenue}</td>
                       <td>{request?.adr}</td>
-                      <td>{request?.revipar}</td>
+                      <td>{request?.revpar}</td>
                       <td>{request?.taxes}</td>
                     </tr>
                   );
                 })}
                 <tr className=" border-b font-semibold">
                   <td>Total</td>
-                  <td>6127</td>
-                  <td>6127</td>
-                  <td>6048</td>
                   <td></td>
-                  <td>N500,117,189.72</td>
-                  <td>0</td>
+                  <td>{reportSummary?.total_nights_booked}</td>
+                  <td>{reportSummary?.total_bookings}</td>
                   <td></td>
                   <td></td>
-                  <td>N90117,189.72</td>
+                  <td></td>
+                  <td></td>
+                  <td></td>
+                  <td>{reportSummary?.total_revenue}</td>
                 </tr>
               </tbody>
             </table>
@@ -112,7 +138,7 @@ export default function DailyRoomReportTable() {
                 {
                   id: 1,
                   label: "Arriving",
-                  value: 12,
+                  value: 0,
                 },
                 {
                   id: 2,
@@ -122,7 +148,7 @@ export default function DailyRoomReportTable() {
                 {
                   id: 3,
                   label: "Stayover",
-                  value: 131,
+                  value: 0,
                 },
                 {
                   id: 4,